% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intoOne.R
\name{intoOne}
\alias{intoOne}
\title{intoOne Function}
\usage{
intoOne(name, path)
}
\arguments{
\item{name}{A string for the name of questionnaire.}

\item{path}{A string for the path of dataset file.}
}
\value{
An object with both raw and scored data.
}
\description{
This function is some very simple code that put all the datafiles in one dataset.
}
\examples{
 \dontrun{
    BBSIQ <- intoOne('BBSIQ','~/active_data/')
   }
 \dontrun{
    intoOne('All','~/active_data/')
   }
}
\keyword{pre-analysis}

