% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBSIQScore.R
\name{score.BBSIQ}
\alias{score.BBSIQ}
\title{score.BBSIQ Function}
\usage{
score.BBSIQ(data)
}
\arguments{
\item{data}{A data object}
}
\value{
To Be Specified.
}
\description{
Scoring method for BBSIQ object.
}
\details{
This is a small function to help score the PI object more effectively.
}
\keyword{datasets}
\keyword{socre}

