% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take.R
\name{take}
\alias{take}
\title{take Function}
\usage{
take(path, type, multiple = FALSE)
}
\arguments{
\item{path}{A string for the path of dataset file.}

\item{type}{A variable indicating the data stage, default to be raw data.}

\item{multiple}{A Boolean variable indicating whether this questionnaire has been used in multiple sessions.}
}
\value{
An object with both raw and scored data.
}
\description{
Import the dataset file, attribute the class of the dataset, and then score it.
}
\details{
The function is used to read and score dataset.
}
\examples{
 \dontrun{
     BBSIQ <- take("BBSIQ.csv","BBSIQ",TRUE)
   }
}
\keyword{import}
\keyword{read}

